# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vehicle Sale and Purchase Platform with Online Auction System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: getVehicles
      tags: [vehicles]
      summary: List all approved vehicles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: sellerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleListResponse"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a new vehicle listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vehicles/{vehicleId}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle by ID
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle listing
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleBody"
      responses:
        "200":
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete vehicle listing
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vehicles/{vehicleId}/buy-now:
    post:
      operationId: buyNow
      tags: [vehicles]
      summary: Buy vehicle immediately at buy-now price
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: No buy-now price or vehicle not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        title:
          type: string
        description:
          type: string
        price:
          type: number
        buyNowPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: ["integer", "null"]
        condition:
          type: string
          enum: [new, used, certified_pre_owned]
        status:
          type: string
          enum: [pending, approved, rejected, sold]
        hasActiveAuction:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sellerId
        - sellerName
        - title
        - description
        - price
        - images
        - make
        - model
        - year
        - condition
        - status
        - hasActiveAuction
        - createdAt
    VehicleListResponse:
      type: object
      properties:
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
        total:
          type: integer
      required:
        - vehicles
        - total
    CreateVehicleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        buyNowPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        mileage:
          type: ["integer", "null"]
        condition:
          type: string
          enum: [new, used, certified_pre_owned]
      required:
        - title
        - description
        - price
        - images
        - make
        - model
        - year
        - condition
    UpdateVehicleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        buyNowPrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        mileage:
          type: ["integer", "null"]
        condition:
          type: string
          enum: [new, used, certified_pre_owned]
...