# lib/api-zod/src/generated/types/createVehicleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Sale and Purchase Platform with Online Auction System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateVehicleBodyCondition } from "./createVehicleBodyCondition";

export interface CreateVehicleBody {
  title: string;
  description: string;
  price: number;
  /** @nullable */
  buyNowPrice?: number | null;
  images: string[];
  make: string;
  model: string;
  year: number;
  /** @nullable */
  mileage?: number | null;
  condition: CreateVehicleBodyCondition;
}